Rename to init and add utils
