+++ internal/super/init_utils.go
@@ -1,0 +1,136 @@
+/*
+ * Flow CLI
+ *
+ * Copyright Flow Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package super
+
+import (
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+)
+
+// getReadmeFileName returns the appropriate README filename, avoiding conflicts
+func getReadmeFileName(targetDir string) string {
+	if _, err := os.Stat(filepath.Join(targetDir, defaultReadmeFile)); err == nil {
+		return flowReadmeFile
+	}
+	return defaultReadmeFile
+}
+
+// copyDirContents copies all files and directories from src to dst
+func copyDirContents(src, dst string) error {
+	srcInfo, err := os.Stat(src)
+	if err != nil {
+		return err
+	}
+
+	if !srcInfo.IsDir() {
+		return fmt.Errorf("source is not a directory")
+	}
+
+	// Read all entries in the source directory
+	entries, err := os.ReadDir(src)
+	if err != nil {
+		return err
+	}
+
+	// Copy each entry
+	for _, entry := range entries {
+		srcPath := filepath.Join(src, entry.Name())
+		dstPath := filepath.Join(dst, entry.Name())
+
+		if entry.IsDir() {
+			// Create directory and recursively copy its contents
+			err := os.MkdirAll(dstPath, defaultDirPerm)
+			if err != nil {
+				return err
+			}
+			err = copyDirContents(srcPath, dstPath)
+			if err != nil {
+				return err
+			}
+		} else {
+			// Copy file
+			err := copyFile(srcPath, dstPath)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+// copyFile copies a single file from src to dst
+func copyFile(src, dst string) error {
+	srcFile, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer srcFile.Close()
+
+	dstFile, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	defer dstFile.Close()
+
+	_, err = io.Copy(dstFile, srcFile)
+	if err != nil {
+		return err
+	}
+
+	// Copy file permissions
+	srcInfo, err := srcFile.Stat()
+	if err != nil {
+		return err
+	}
+	return os.Chmod(dst, srcInfo.Mode())
+}
+
+// getTargetDirectory checks if the specified directory path is suitable for use.
+// It verifies that the path points to an existing, empty directory.
+// If the directory does not exist, the function returns the path without error,
+// indicating that the path is available for use (assuming creation is handled elsewhere).
+func getTargetDirectory(directory string) (string, error) {
+	pwd, err := os.Getwd()
+	if err != nil {
+		return "", err
+	}
+
+	target := filepath.Join(pwd, directory)
+	info, err := os.Stat(target)
+	if !os.IsNotExist(err) {
+		if !info.IsDir() {
+			return "", fmt.Errorf("%s is a file", target)
+		}
+
+		file, err := os.Open(target)
+		if err != nil {
+			return "", err
+		}
+		defer file.Close()
+
+		_, err = file.Readdirnames(1)
+		if err != io.EOF {
+			return "", fmt.Errorf("directory is not empty: %s", target)
+		}
+	}
+	return target, nil
+}
